+++ src/lambdaisland/launchpad/deps.clj
@@ -3,7 +3,9 @@
    [clojure.edn :as edn]
    [clojure.java.io :as io]
    [clojure.tools.deps.alpha :as deps]
-   [lambdaisland.classpath.watch-deps :as watch-deps]))
+   [clojure.string :as str]
+   [lambdaisland.classpath.watch-deps :as watch-deps]
+   [lambdaisland.classpath :as licp]))
 
 (defn basis [opts]
   (let [deps-local-file (io/file "deps.local.edn")
@@ -13,6 +15,22 @@
     (deps/create-basis (-> opts
                            (update :aliases concat (:launchpad/aliases deps-local))
                            (assoc :extra extra-deps)))))
+
+(defn write-cpcache-file
+  "Write the current classpath to well-known location. Can be used to configure
+  things like LSP that try to guess the classpath without having a running
+  process.
+
+  e.g. .lsp/config.edn
+  ```
+  {:project-specs [:classpath-cmd [\"cat\" \".cpcache/launchpad.edn\"]]}
+  ```
+  "
+  []
+  (spit ".cpcache/launchpad.cp"
+        (str/join
+         ":"
+         (time (mapcat second (licp/classpath-chain))))))
 
 (defn watch-handlers [opts]
   (let [basis (basis opts)
@@ -27,7 +45,9 @@
                      (conj (watch-deps/canonical-path (:extra opts)))
                      :always
                      (concat (:watch-paths opts)))
-        handler (partial #'watch-deps/on-event deps-paths opts)]
+        handler (fn [e]
+                  (#'watch-deps/on-event deps-paths opts e)
+                  (write-cpcache-file))]
     (into {}
           (map (fn [p]
                  [(str p) handler]))