+++ src/lambdaisland/launchpad/shadow.clj
@@ -106,5 +106,4 @@
           (recur))))))
 
 #_(start-builds :main)
-
-(require 'shadow.build)
+#_(require 'shadow.build)